Integrated end to end metrics display
The metrics calculation likely needs an overhaul, but the signal is
piped through from Pipeline to frontend.

What's wrong at the moment:
* ISO 8601 timestamp formulation is off. The time format looks wrong.
* Number weighting is wrong. The execution counts and processed
candidates are not correct yet and need to be divided by the time
elapsed since reporting.

+++ include/beast/pipeline.hpp
@@ -59,9 +59,9 @@ class Pipeline {
    * @brief Holds metrics describing the performance of an individual pipe
    */
   struct PipeMetrics {
-    uint32_t execution_count;
-    std::unordered_map<uint32_t, uint32_t> inputs_received;
-    std::unordered_map<uint32_t, uint32_t> outputs_sent;
+    double execution_count;
+    std::unordered_map<uint32_t, double> inputs_received;
+    std::unordered_map<uint32_t, double> outputs_sent;
   };
 
   /**
